NEWS: Include description for release 0.7.3
The description for this release, unlike all other descriptions,
was inside a <p> element; however, the XSLT stylesheet contains a
template that drops all <p> elements from the output file, so it
never made it to the generated NEWS file.

Use a <li> element, same as all other releases, instead.