NEWS: Split old entries (2008)
The changes for releases earlier than 0.7.1 were mostly lumped
together as opposed to being tidly organized with one change per
line, like we have done from that point onwards.

As a result, they look awful in the HTML version and don't work
too well in the plain text version either.

Luckily, except for the very first releases, the information is
still very detailed, so it's enough to organize it properly.